+++ vite.config.js
@@ -4,8 +4,10 @@ import react from '@vitejs/plugin-react'
 export default defineConfig({
   plugins: [react()],
   server: {
+    port: 9527,
+    strictPort: true,
     proxy: {
-      '/api': 'http://localhost:3001'
+      '/api': 'http://localhost:4321'
     }
   }
 })